+++ tests/testthat/test-quick-get.R
@@ -1,13 +1,14 @@
-# test_that("spod_quick_get_od fails out of range dates", {
-#   skip_on_ci()
-#   skip_on_cran()
-#   expect_error(
-#     spod_quick_get_od(
-#       date = "2021-12-31",
-#     ),
-#     ".*Must be within valid range.*"
-#   )
-# })
+test_that("spod_quick_get_od fails out of range dates", {
+  skip_on_ci()
+  skip_on_cran()
+  expect_error(
+    spod_quick_get_od(
+      date = "2021-12-31",
+    ),
+    ".*Must be within valid range.*"
+  )
+})
+
 
 test_that("spod_quick_get_od fails on invalid date format", {
   expect_error(
@@ -44,7 +45,7 @@ test_that("spod_quick_get_od fails on invalid municipality IDs", {
       date = "2022-01-03",
       id_origin = "invalid"
     ),
-    ".*Invalid municipality IDs detected.*"
+    ".*Invalid municipality ID.*"
   )
 
   expect_error(
@@ -52,7 +53,7 @@ test_that("spod_quick_get_od fails on invalid municipality IDs", {
       date = "2022-01-04",
       id_destination = "invalid"
     ),
-    ".*Invalid municipality IDs detected.*"
+    ".*Invalid municipality ID.*"
   )
 
   expect_error(
@@ -61,6 +62,6 @@ test_that("spod_quick_get_od fails on invalid municipality IDs", {
       id_origin = "invalid",
       id_destination = "invalid"
     ),
-    ".*Invalid municipality IDs detected.*"
+    ".*Invalid municipality ID.*"
   )
 })